Add comment policy and API error message
Add CommentPolicy for API endpoint auth.
Return error message instead of whole exception.

+++ app/Http/Controllers/Api/V1/PostCommentController.php
@@ -78,6 +78,8 @@ class PostCommentController extends Controller
      */
     public function update(BlogPost $post, Comment $comment, StoreComment $request)
     {
+        $this->authorize($comment);
+
         $comment->content = $request->input('content');
         $comment->save();
 
@@ -92,6 +94,8 @@ class PostCommentController extends Controller
      */
     public function destroy(BlogPost $post, Comment $comment)
     {
+        $this->authorize($comment);
+
         $comment->delete();
 
         return response()->noContent();
